% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wmesaclient.R
\name{absPosition.mesa}
\alias{absPosition.mesa}
\title{Posição no sistema de referência abosluto}
\usage{
\method{absPosition}{mesa}(dev)
}
\arguments{
\item{dev}{Um objeto de classe \code{mesa} contendo informações sobre servidor XML-RPC}
}
\value{
Lista contendo o ângulo da mesa
}
\description{
Retorna a posição da mesa giratória no sistema de referência absoluto.
O sistema de referência absoluto é o que está na placa controladora da mesa giratória
}
\examples{
absPosition(dev)

}
