% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wmesaclient.R
\name{mesaDev}
\alias{mesaDev}
\title{Criar conexão com a mesa giratório do túnel de vento do IPT}
\usage{
mesaDev(url = "localhost", port = 9596)
}
\arguments{
\item{url}{String com URL do servidor XML-RPC}

\item{port}{Inteiro com o número da porta TCP/IP usado pelo XML-RPC}
}
\value{
Um objeto de classe \code{mesa}
}
\description{
Cria um objeto de classe \code{mesa} que controla a comunicação com
o servidor XML-RPC da mesa giratória do túnel de vento do IPT.
}
\examples{
dev <- mesaDev()

}
