% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wmesaclient.R
\name{move.mesa}
\alias{move.mesa}
\title{Gira a mesa do túnel de vento do IPT.}
\usage{
\method{move}{mesa}(dev, x, a = FALSE, r = FALSE, sync = FALSE)
}
\arguments{
\item{dev}{Um objeto de classe \code{mesa} contendo informações sobre servidor XML-RPC}

\item{x}{Número informando o ângulo em graus que se deseja girar}

\item{a}{TRUE/FALSE, informa se o movimento se dá na referência absoluta ou relativa}

\item{r}{TRUE/FALSE, movimento relativo em relação ao ponto atual?}

\item{@sync}{TRUE/FALSE Se sincronizado, a função só retorna depois de executado o movimento}
}
\description{
Gira a mesa do túnel de vento do ipt. Valores positivos do ângulo de rotação
resultam em movimento no sentido anti-horário.
}
\examples{
move(dev, 5, r=TRUE)

}
