% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wmesaclient.R
\name{position.mesa}
\alias{position.mesa}
\title{Posição da mesa giratória}
\usage{
\method{position}{mesa}(dev)
}
\arguments{
\item{dev}{Um objeto de classe \code{mesa} contendo informações sobre servidor XML-RPC}
}
\value{
Lista contendo o ângulo
}
\description{
Retorna a posição da mesa giratória no sistema de referência relativo.
O sistema de referência absoluto é o que está na placa controladora da mesa giratória.
}
\examples{
position(dev)

}
