% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wmesaclient.R
\name{rmove.mesa}
\alias{rmove.mesa}
\title{Movimento relativo da mesa giratória}
\usage{
\method{rmove}{mesa}(dev, x = "", y = "", z = "", sync = FALSE)
}
\arguments{
\item{dev}{Um objeto de classe \code{mesa} contendo informações sobre servidor XML-RPC}

\item{x}{Número informando o ângulo que se deseja girar a mesa}

\item{@sync}{TRUE/FALSE Se sincronizado, a função só retorna depois de executado o movimento}
}
\description{
Gira a mesa em relação à posição atual. Valores positivos de giro correspondem ao
sentido anti-horário.
}
\examples{
rmove(dev, 5)

}
