% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wmesaclient.R
\name{setReference.mesa}
\alias{setReference.mesa}
\title{Cria novo sistema de referência}
\usage{
\method{setReference}{mesa}(dev, xref = 0)
}
\arguments{
\item{dev}{Um objeto de classe \code{mesa} contendo informações sobre servidor XML-RPC}

\item{xref}{Valor que a coordenada x atual toma.}
}
\description{
Configura um novo sistema de referência onde a posição atual da mesa giratória tem
valore \code{xref}. Este sistema de referência relativo
tem a mesma orientação que o sistema de placa controladora. O que muda é a origem.
}
\examples{
setReference(dev, 100)

}
