% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wmesaclient.R
\name{stopnow.mesa}
\alias{stopnow.mesa}
\title{Paramada imediata}
\usage{
\method{stopnow}{mesa}(dev)
}
\arguments{
\item{dev}{Um objeto de classe \code{mesa} contendo informações sobre servidor XML-RPC}
}
\description{
Envia comando para a mesa giratória para qualquer movimento.
}
\examples{
move(dev, 30)
stopnow(dev)

}
